Moved user snippets meta logic to CodesplainAppBar component
