+++ src/containers/CodesplainAppBar.jsx
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import AppBar from 'material-ui/AppBar';
 import { browserHistory } from 'react-router';
 import cookie from 'react-cookie';
+import { saveUserSnippets } from '../actions/app';
 
 import LoginButton from '../components/buttons/LoginButton'
 import AppMenu from '../components/menus/AppMenu';
@@ -31,6 +32,13 @@ class CodesplainAppBar extends React.Component {
     this.handleSignOut = this.handleSignOut.bind(this);
     this.handleTitleClick = this.handleTitleClick.bind(this);
     this.onLoginClick = this.onLoginClick.bind(this);
+    this.fetchUserSnippets = this.fetchUserSnippets.bind(this);
+  }
+
+  componentDidMount() {
+    const token = cookie.load('token');
+    const username = cookie.load('username');
+    this.fetchUserSnippets(token, username);
   }
 
   handleSignOut() {
@@ -50,8 +58,6 @@ class CodesplainAppBar extends React.Component {
     // later when GitHub sends the user back to the app, and send the user back
     // to the page they were on when they clicked the login button.
     const appState = JSON.stringify(this.props.appState);
-    cookie.remove('signInState', { path: '/' });
-    cookie.remove('signInRedirect', { path: '/' });
     cookie.save('signInState', appState, { path: '/' });
     cookie.save('signInRedirect', location.pathname, { path: '/' });
 
@@ -60,30 +66,50 @@ class CodesplainAppBar extends React.Component {
     window.location = GITHUB_URL;
   }
 
+  // Get meta data about the user's snippets and save to redux
+  async fetchUserSnippets(token, username) {
+    const { dispatch } = this.props;
+    // let res;
+    // try {
+    //   const headers = {
+    //     Accept: 'application/json',
+    //     Authorization: `token ${token}`,
+    //   };
+    //   res = await axios.get(`${API_URL}/users/${username}/snippets`, { headers });
+    // } catch(e) {
+    //   return;
+    // }
+    const snippetMeta = [
+      {
+        snippetName: 'Test Snippet 1',
+        language:    'Python 3',
+        lastEdited:  '2017-04-05T12:52:20.099Z',
+        'private':   true,
+      },
+      {
+        snippetName: 'Test Snippet 2',
+        language:    'Python 3',
+        lastEdited:  '2017-04-05T12:52:20.099Z',
+        'private':   true,
+      }
+    ];
+
+    dispatch(saveUserSnippets(snippetMeta));
+  }
+
   render() {
-    let snippetTitles = [];
-    if (this.props.appState.snippetMeta)
-      snippetTitles = this.props.appState.snippetMeta.map(s => s.snippetName);
+    const { snippetMeta } = this.props.appState;
+    const titles = snippetMeta ? snippetMeta.map(s => s.snippetName) : null;
     const rightElement = this.state.isLoggedIn ?
       <AppMenu
         onSignOut={this.handleSignOut}
-        // snippetTitles={['none']}
-        snippetTitles={snippetTitles}
+        snippetTitles={titles}
         onTitleClicked={this.handleTitleClick}
-      /> :
+      />
+      :
       <LoginButton
         onClick={this.onLoginClick}
       />;
-    console.log(this.props.appState)
-    // let snippetTitles = [];
-    // if (this.props.appState.snippetMeta)
-    //   snippetTitles = this.props.appState.snippetMeta.map(s => s.snippetName);
-    // const rightElement =
-    // <AppMenu
-    //   onSignOut={this.handleSignOut}
-    //   snippetTitles={snippetTitles}
-    //   onTitleClicked={this.handleTitleClick}
-    // />;
     return (
       <AppBar
         showMenuIconButton={false}
